Allow for a `callback` sub
I didn't know of a better way to run extra tests once we connect to the
server. Maybe the diagnostic wants to check if a file exists on the
server. Maybe the diagnostic wants to check if it can push a file to the
server. The possibilities are endless, so I just figured to allow the
developer the power to come up with their own if they want.